Wire up SigningService
Changes the Ethereum gRPC service to make calls to an inner
`SigningService` which holds the `signing::Keyring`, rather than letting
the Ethereum service hold the keyring directly.

This is futureproofing for when we have more than one gRPC service that
might want to access the signing keyring, e.g. Tendermint "privval".